+++ api/pkg.ts
@@ -2,11 +2,10 @@ import { NowRequest, NowResponse } from "@now/node";
 import { pkg } from "./_utils/pkg";
 
 export default async (request: NowRequest, response: NowResponse) => {
-  const requestUrl = request.url;
-
   await pkg({
     query: request.query,
-    requestUrl,
+    requestUrl: request.url,
+    parseFromUrl: false,
     response,
   });
 };
